requirements.txt
alembic==1.6.5
appdirs==1.4.4
asgiref==3.4.1
asyncpg==0.24.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
black==21.7b0
cffi==1.14.6
cfgv==3.3.0
click==8.0.1
cohesion==1.0.0
databases==0.4.3
distlib==0.3.2
fastapi==0.68.0
filelock==3.0.12
flake8==3.9.2
greenlet==1.1.1
h11==0.12.0
identify==2.2.13
loguru==0.5.3
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
passlib==1.7.4
pathspec==0.9.0
pep517==0.11.0
pip-tools==6.2.0
platformdirs==2.2.0
pre-commit==2.14.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
python-editor==1.0.4
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
uvicorn==0.15.0
virtualenv==20.7.2
